Add files via upload
Left mouse to select new x/y location which influences a lot. I used interpolation to make the transitions smooth. Right mouse turns on "x-ray mode", and center mouse button saves an image. Press s to turn the spiral off and on (it comes back out slowly), space bar to turn the stroke on the circles off and on, and the arrow keys to shift the circle stroke location and range (based on the recursion level). The location influences the stroke weight, and also, the mouse can go way off screen if you drag out of bounds. These can make some interesting variations.